requirements.txt
discord.py==2.4.0
groq
edge-tts
python-dotenv
